requirements.txt
Flask==2.1.2
flask-restx==0.5.1
Werkzeug==2.1.2
gunicorn==20.1.0
pip~=23.2.1
attrs~=22.1.0
cryptography~=38.0.3
tornado~=6.2
Jinja2~=3.1.2
pytz~=2020.1
MarkupSafe~=2.1.1
click~=8.1.7
python-dotenv~=0.19.0
itsdangerous~=2.1.2
aniso8601~=9.0.1
jsonschema~=4.17.3
idna~=3.4
webcolors~=1.11.1
setuptools~=47.1.0
packaging~=21.3
zipp~=3.8.1
six~=1.16.0
pytest~=7.2.0
pyserial~=3.5